Drop the bigdecimal crate

+++ models/src/avro_gen/models.rs
@@ -1,6 +1,6 @@
-use serde::{Deserialize, Serialize};
-use lazy_static;
 use avro_rs::schema::Schema;
+use lazy_static;
+use serde::{Deserialize, Serialize};
 
 #[derive(Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Clone, Deserialize, Serialize)]
 pub enum TradeType {
@@ -9,7 +9,6 @@ pub enum TradeType {
     #[serde(rename = "SELL")]
     Sell,
 }
-
 
 lazy_static! {
     pub static ref LIVETRADE_SCHEMA : Schema = Schema::parse_str("{\"type\":\"record\",\"name\":\"LiveTrade\",\"fields\":[{\"name\":\"event_ms\",\"type\":\"long\"},{\"name\":\"pair\",\"type\":\"string\"},{\"name\":\"amount\",\"type\":\"float\"},{\"name\":\"price\",\"type\":\"float\"},{\"name\":\"tt\",\"type\":\"int\"}]}").unwrap();
@@ -20,8 +19,8 @@ lazy_static! {
 pub struct LiveTrade {
     pub event_ms: i64,
     pub pair: String,
-    pub amount: f32,
-    pub price: f32,
+    pub amount: f64,
+    pub price: f64,
     pub tt: i32,
 }
 
@@ -37,7 +36,6 @@ impl Default for LiveTrade {
     }
 }
 
-
 lazy_static! {
     pub static ref LIVEORDER_SCHEMA : Schema = Schema::parse_str("{\"type\":\"record\",\"name\":\"LiveOrder\",\"fields\":[{\"name\":\"event_ms\",\"type\":\"long\"},{\"name\":\"amount\",\"type\":\"float\"},{\"name\":\"pair\",\"type\":\"string\"},{\"name\":\"price\",\"type\":\"float\"},{\"name\":\"tt\",\"type\":\"int\"}]}").unwrap();
 }
@@ -46,9 +44,9 @@ lazy_static! {
 #[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
 pub struct LiveOrder {
     pub event_ms: i64,
-    pub amount: f32,
+    pub amount: f64,
     pub pair: String,
-    pub price: f32,
+    pub price: f64,
     pub tt: i32,
 }
 
@@ -64,7 +62,6 @@ impl Default for LiveOrder {
     }
 }
 
-
 lazy_static! {
     pub static ref ORDERBOOK_SCHEMA : Schema = Schema::parse_str("{\"type\":\"record\",\"name\":\"Orderbook\",\"fields\":[{\"name\":\"event_ms\",\"type\":\"long\"},{\"name\":\"pair\",\"type\":\"string\"},{\"name\":\"asks\",\"type\":{\"type\":\"array\",\"items\":{\"type\":\"array\",\"items\":\"float\"}}},{\"name\":\"bids\",\"type\":{\"type\":\"array\",\"items\":{\"type\":\"array\",\"items\":\"float\"}}}]}").unwrap();
 }
@@ -74,8 +71,8 @@ lazy_static! {
 pub struct Orderbook {
     pub event_ms: i64,
     pub pair: String,
-    pub asks: Vec<Vec<f32>>,
-    pub bids: Vec<Vec<f32>>,
+    pub asks: Vec<Vec<f64>>,
+    pub bids: Vec<Vec<f64>>,
 }
 
 impl Default for Orderbook {
